Moved existing test files to proper packages.
We should put the test files in the same package as the source code they are testing, just under /test instead of /app. This allows us to not have to import the classes we are testing and gives us access to package private fields.
